requirements.txt
matplotlib==3.7.3
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.1
Pillow==10.0.1
scipy==1.11.2
skimage==0.0
